Added support for domain authentication
Engaged flask_ldap3_login to support authenticating users against
an LDAP directory, such as a Windows domain's Actve Directory.  Use
of this facility is controlled by instance configuration, and disabled
by default.

Also added a config option to control treating usernames entered on
the login form as case-insensitive. When that is turned on, usernames
are converted to lowercase for the application's internal use.

And added debug logging of the instance configuration.