Fixed the booking calendar sometimes showing no events (#88)
### Fixes: 

* Fixed the booking calendar sometimes showing no events.

Sometimes the booking calendar would show no events despite successfully
loading some from the event source.  This appears to have been a timing
issue of some kind, but dependent on the behavior or the registered
eventSourceSuccess callback.  This change to the callback seems to
resolve the issue with no change to the calendar's observable
behavior.

And the bit removed from the callback shouldn't have been done by
that callback anyway.

* Added an extension point for general app setup modeled on the other extension points.

* Fixed the load_module() function in the app initialization

* It was naming some of the extension modules incorrectly when loading them.